requirements.txt
gql~=3.4.0
python-dotenv